* osx-test:
  NFS and docker host is now working correctly on OSX
  Mysql original image with UID/GID wrapper script. Should make OSX users able to use local data storage
  Reverted: Switched MySQL container.
  Removed port forwarding
  Switched mysql container for use on osx
  Disabled NFS for OSX

+++ dockerfiles/d7/bootstrap/run-drupal.sh
@@ -10,18 +10,21 @@ cp /bootstrap/varnish-default.vcl /etc/varnish/default.vcl
 # Echo's mail into /var/log/mail
 ln -s /bootstrap/sendmail /bin/sendmail
 
-# ===================== Give host a hostname ============
+# ===================== Give host a hostname unless it's already known (on a mac) ============
 SED="$(which sed)"
 NETSTAT="$(which netstat)"
 GREP="$(which grep)"
 AWK="$(which awk)"
 CAT="$(which cat)"
 
-$SED '/dockerhost$/d' /etc/hosts > /etc/hosts.tmp
-DOCKERHOST="$($NETSTAT -nr | $GREP '^0\.0\.0\.0' | $AWK '{print $2}')"
-echo "$DOCKERHOST dockerhost" >> /etc/hosts.tmp
-$CAT /etc/hosts.tmp > /etc/hosts
-rm -rf /etc/hosts.tmp
+host_known=$($CAT /etc/hosts | grep dockerhost)
+if [ "${host_known}" = "" ]; then
+  $SED '/dockerhost$/d' /etc/hosts > /etc/hosts.tmp
+  DOCKERHOST="$($NETSTAT -nr | $GREP '^0\.0\.0\.0' | $AWK '{print $2}')"
+  echo "$DOCKERHOST dockerhost" >> /etc/hosts.tmp
+  $CAT /etc/hosts.tmp > /etc/hosts
+  rm -rf /etc/hosts.tmp
+fi
 # ==
 
 # ==================== Set apache vhost, if necessary ==
